requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-filter==2.4.0